Many code cleanups
* use c++ casts
* systguidize
* fix warnings
* rework qInfo() and friends
* use QStringLiteral

+++ src/actuasense/actuasense_p.h
@@ -33,7 +33,7 @@ public:
     bool m_bDemoError;
 
     int m_iTimeoutMs;
-    int m_iMsSinceLastSet;
+    qint64 m_iMsSinceLastSet;
     bool m_bInStateDesired;
 
     QString m_strOK;
